JGit.FileUtils has trouble with deleting non-existing directories. Win results: Tests run: 9, Failures: 2, Errors: 4, Skipped 0

+++ maven-scm-providers/maven-scm-providers-git/maven-scm-provider-jgit/src/test/java/org/apache/maven/scm/provider/git/jgit/command/changelog/JGitChangeLogCommandTckTest.java
@@ -45,6 +45,9 @@ public class JGitChangeLogCommandTckTest
     protected void deleteDirectory( File directory )
         throws IOException
     {
-        FileUtils.delete( directory, FileUtils.RECURSIVE | FileUtils.RETRY );
+        if ( directory.exists() )
+        {
+            FileUtils.delete( directory, FileUtils.RECURSIVE | FileUtils.RETRY );
+        }
     }
 }
